requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
yfinance>=0.2.36
exchange-calendars>=4.5.0
apscheduler>=3.10.4
aiosqlite>=0.19.0
pyserial>=3.5
tradernet-sdk>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
empyrical-reloaded>=0.5.10
pandas-ta>=0.3.14b
pyfolio-reloaded>=0.9.8
PyPortfolioOpt>=1.5.4
scikit-learn>=1.3.0  # Required for Ledoit-Wolf shrinkage in PyPortfolioOpt
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code quality tools
black>=23.12.0
flake8>=7.0.0
isort>=5.13.0
mypy>=1.8.0
types-requests>=2.31.0
pre-commit>=3.6.0

# Security tools (critical for financial applications)
bandit>=1.7.5
safety>=2.3.5

# Additional code quality
pydocstyle>=6.3.0
coverage>=7.3.0
vulture>=2.11

# Router Bridge client for Arduino Uno Q
msgpack>=1.0.0
tomli>=2.0.0

# gRPC for microservices
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
PyYAML>=6.0.0
